getNearestIndex util for Slider
